Update HelloGit.py
Would it be worth showing what the result would be after running the code, especially for the later matrix parts? It might be cumbersome trying to go between the code/comments and the result after running it. Having both there at the same time would make comparing it easier.
Also, sorry if this ended up in the wrong place. It's my first time editing something.